+++ src/products/controller/products.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ProductsController } from './products.controller';
-import { ProductsService } from './products.service';
+import { ProductsService } from '../service/products.service';
 
 describe('ProductsController', () => {
   let controller: ProductsController;
