+++ src/products/products.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ProductsService } from './products.service';
-import { ProductsController } from './products.controller';
+import { ProductsService } from './service/products.service';
+import { ProductsController } from './controller/products.controller';
 import { ProductsRpository } from './repository/products.repository';
 import { PrismaModule } from '../prisma/prisma.module';
