+++ src/products/service/products.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
-import { CreateProductDto } from './dto/create-product.dto';
-import { UpdateProductDto } from './dto/update-product.dto';
-import { ProductsRpository } from './repository/products.repository'; 
+import { CreateProductDto } from '../dto/create-product.dto';
+import { UpdateProductDto } from '../dto/update-product.dto';
+import { ProductsRpository } from '../repository/products.repository'; 
 
 @Injectable()
 export class ProductsService {
